fix(coin:xrp): iterate over pages passing the token
