+++ .changeset/yellow-coins-agree.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-xrp": patch
+---
+
+Fix iteration on api/listOperations
